+++ Packages/com.nekuzaky.input-prompts/Editor/src/InputPromptSettings.cs
@@ -21,6 +21,7 @@ namespace Nekuzaky.InputPrompts.Editor
         public bool m_outlineIcons;
         public bool m_coloredFaceButtons = true;
         public bool m_applyTextureSettings = true;
+        public bool m_generateSpriteAssets = true;
 
         public InputDeviceStyle m_gamepadFallbackStyle = InputDeviceStyle.Xbox;
         public InputDeviceStyle m_defaultStyle = InputDeviceStyle.KeyboardMouse;
@@ -45,6 +46,9 @@ namespace Nekuzaky.InputPrompts.Editor
 
         public string SetPathFor(InputDeviceStyle style) =>
             $"{m_outputFolder}/{InputPromptGenerator.SetPrefix}{style}.asset";
+
+        public string SpriteAssetPathFor(InputDeviceStyle style) =>
+            $"{m_outputFolder}/{InputPromptGenerator.SpriteAssetPrefix}{style}.asset";
 
         public bool IsStyleEnabled(InputDeviceStyle style) => !m_disabledStyles.Contains(style);